Changed an int to size_t and added more friends to input.txt

+++ main.cpp
@@ -24,7 +24,7 @@ int main(){
 		string friend1 = "";
 		while(getline(file,line)){
 			int position = 0;
-			for(int i = 0; i < line.length(); i++){
+			for(size_t i = 0; i < line.length(); i++){
 				if(line[i] == '\n')
 					break;
 				if(line[i] == ',')
